feat(qa): 検索機能を強化し、IME対応の入力処理を統合
