+++ src/pages/qa/_QaSearch.tsx
@@ -1,16 +1,60 @@
 import SearchResultItem from "./_QaSearchResultItem";
 import type { QaSearchItem } from "./_qa";
-import { useImeAwareInput } from "./_useImeAwareInput";
-import { useQaSearch } from "./_useQaSearch";
+import { ensureNotNullish, UnreachableError } from "@/helper";
 import { faMagnifyingGlass, faXmark } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import Fuse from "fuse.js";
+import type { FuseResultMatch, IFuseOptions, RangeTuple } from "fuse.js";
+import {
+  useMemo,
+  useState,
+  type ChangeEventHandler,
+  type CompositionEventHandler,
+} from "react";
 
 type QaSearchProps = {
   items: QaSearchItem[];
 };
 
+type SearchKey = "category" | "question" | "answer";
+
+export type QaSearchResult = {
+  item: QaSearchItem;
+  indicesByKey: Record<SearchKey, readonly RangeTuple[]>;
+};
+
+type SearchResultState =
+  | { kind: "idle" }
+  | { kind: "empty" }
+  | { kind: "matched"; results: QaSearchResult[] };
+
+type ImeAwareInputState = {
+  input: string;
+  committed: string;
+};
+
+type ImeAwareInput = {
+  input: string;
+  committed: string;
+  onChange: ChangeEventHandler<HTMLInputElement>;
+  onCompositionEnd: CompositionEventHandler<HTMLInputElement>;
+  clear: () => void;
+};
+
 const SEARCH_INPUT_ID = "qa-search-input";
 const PAGE_TITLE_ID = "qa-page-title";
+const MAX_RESULTS = 12;
+const FUSE_OPTIONS = {
+  keys: [
+    { name: "question", weight: 0.55 },
+    { name: "answer", weight: 0.35 },
+    { name: "category", weight: 0.1 },
+  ],
+  includeMatches: true,
+  includeScore: true,
+  ignoreLocation: true,
+  threshold: 0.2,
+} satisfies IFuseOptions<QaSearchItem>;
 
 export default function QaSearch({ items }: QaSearchProps) {
   const { input, committed, onChange, onCompositionEnd, clear } =
@@ -96,4 +140,86 @@ export default function QaSearch({ items }: QaSearchProps) {
       )}
     </section>
   );
+}
+
+function useImeAwareInput(initial: string): ImeAwareInput {
+  const [state, setState] = useState<ImeAwareInputState>({
+    input: initial,
+    committed: initial,
+  });
+
+  const onChange: ChangeEventHandler<HTMLInputElement> = (event) => {
+    const value = event.currentTarget.value;
+    if (!(event.nativeEvent instanceof InputEvent)) {
+      throw new UnreachableError();
+    }
+    const isComposing = event.nativeEvent.isComposing;
+    setState((current) => ({
+      input: value,
+      committed: isComposing ? current.committed : value,
+    }));
+  };
+
+  const onCompositionEnd: CompositionEventHandler<HTMLInputElement> = (
+    event,
+  ) => {
+    const value = event.currentTarget.value;
+    setState({ input: value, committed: value });
+  };
+
+  const clear = () => {
+    setState({ input: "", committed: "" });
+  };
+
+  return {
+    input: state.input,
+    committed: state.committed,
+    onChange,
+    onCompositionEnd,
+    clear,
+  };
+}
+
+function useQaSearch(
+  items: QaSearchItem[],
+  query: string,
+): SearchResultState {
+  const fuse = useMemo(() => new Fuse(items, FUSE_OPTIONS), [items]);
+  return useMemo<SearchResultState>(() => {
+    const trimmed = query.trim();
+    if (trimmed.length === 0) {
+      return { kind: "idle" };
+    }
+    const results = fuse.search(trimmed, { limit: MAX_RESULTS }).map(
+      (result): QaSearchResult => ({
+        item: result.item,
+        indicesByKey: buildIndicesByKey(ensureNotNullish(result.matches)),
+      }),
+    );
+    if (results.length === 0) {
+      return { kind: "empty" };
+    }
+    return { kind: "matched", results };
+  }, [fuse, query]);
+}
+
+function buildIndicesByKey(
+  matches: readonly FuseResultMatch[],
+): Record<SearchKey, readonly RangeTuple[]> {
+  const indicesByKey: Record<SearchKey, readonly RangeTuple[]> = {
+    category: [],
+    question: [],
+    answer: [],
+  };
+  for (const match of matches) {
+    indicesByKey[parseSearchKey(match.key)] = match.indices;
+  }
+  return indicesByKey;
+}
+
+function parseSearchKey(key: string | undefined): SearchKey {
+  if (key === "category" || key === "question" || key === "answer") {
+    return key;
+  }
+  throw new UnreachableError();
 }
