+++ src/pages/qa/_QaSearchResultItem.tsx
@@ -1,5 +1,5 @@
+import type { QaSearchResult } from "./_QaSearch";
 import { QUESTION_HEADING_PREFIX } from "./_qa";
-import type { QaSearchResult } from "./_useQaSearch";
 import { assertNonNullable } from "@/helper";
 import type { RangeTuple } from "fuse.js";
 import type { ReactNode } from "react";
